pyproject.toml
[tool.poetry]
name = "gifted"
version = "0.1.0"
description = ""
authors = ["liammcbride <mailmcbride56@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
